Moved scripts to bin folder
