+++ package.json
@@ -13,10 +13,10 @@
     "test:e2e": "jest --testPathPattern=e2e --runInBand",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
-    "db:create-users-table": "ts-node ./src/scripts/create-users-table.ts",
-    "db:delete-users-table": "ts-node ./src/scripts/delete-users-table.ts",
-    "db:put-user": "ts-node ./src/scripts/put-user.ts",
-    "db:get-user-from-sub": "ts-node ./src/scripts/get-user-from-sub.ts",
+    "db:create-users-table": "ts-node bin/dynamodb/create-users-table.ts",
+    "db:delete-users-table": "ts-node bin/dynamodb/delete-users-table.ts",
+    "db:put-user": "ts-node bin/dynamodb/put-user.ts",
+    "db:get-user-from-sub": "ts-node bin/dynamodb/get-user-from-sub.ts",
     "e2e-add-redirect-uri": "ts-node bin/add-redirect-uri-to-connect.ts",
     "e2e-remove-redirect-uri": "ts-node bin/remove-redirect-uri-from-connect.ts"
   },
