+++ src/pages/api/oauth/callback.ts
@@ -49,8 +49,6 @@ const handler: Handler = async (
       id_token,
     };
 
-    console.log(oauthUserInfo);
-
     await getAndPutUser(oauthUserInfo);
 
     request.session.set("user-session-id", documentId);
